add logic mines game

+++ app/src/main/java/com/example/h2bet/ui/game/mines/MinesDataGame.kt
@@ -4,7 +4,7 @@ class MinesDataGame {
     data class MinesModel(
         var id:Int,
         var select: Boolean,
-        var value: Float,
+        var value: Double,
         var type: Boolean
     )
